Changes made for successful deployment of version #2!
Needed to reduce number of feed posts to fetch from 250 down to 50.
Would time out often if it was too large.  Need to figure this out
still.

Also had to remove a JSON library that GAE was complaining about because
apparently it is a JDK8 SDK and GAE only supports JDK7 at the moment.

Updated appengine-web.xml to version 2 and deployed!!